Significantly improved UNC-support
Disabled forced computer search (should be done in a separate thread, if
at all).
Instead add host-entries for explicitely accessed UNC-paths to treeview.
To not do the same work twice the ShellBrowser should use the same
information source, as that would require more effort for now the list
for the 'Network'-node stays empty, the info is in the treeview anyway.