Fix 'do not sort new items' to behave as expected.
And the expected behaviour is to not update sorting because of adding
files or folders or because of renaming them, thats how Explorer behaved
before Windows 7 which then annoyed their users to no end with this
forced sorting nonsense.
Which couldn't be disabled by easy means in Win 7+ Explorer.
That's why this setting is important and needs to work fully - you want
updated sorting, it's just one F5 away.